LayoutTests/fast/forms/select/options-indexed-properties.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<select id="select1">
    <option value="a">A</option>
    <option value="b">B</option>
</select>
<div id="console"></div>

<script>
description("https://bugs.webkit.org/show_bug.cgi?id=85937 - [Chromium][Forms] HTMLOptionsCollection doesn't have indexed properties on property enumeration");
var properties = {};
var select1 = document.getElementById("select1");
var options = select1.options;
for (var propertyKey in options) {
    properties[propertyKey] = options[propertyKey];
}
shouldBeEqualToString('properties[0].value', "a");
shouldBeEqualToString('properties[1].value', "b");
</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
