third_party/blink/web_tests/fast/forms/select/options-indexed-properties-expected.txt

A
B

https://bugs.webkit.org/show_bug.cgi?id=85937 - [Chromium][Forms] HTMLOptionsCollection doesn't have indexed properties on property enumeration

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS properties[0].value is "a"
PASS properties[1].value is "b"
PASS successfullyParsed is true

TEST COMPLETE
